api/requirements.txt
fastapi
gunicorn
pydantic
pymysql
sqlalchemy
starlette
uvicorn
python-decouple
aiofiles
PyJWT
python-multipart